refactor: improve tool package cohesion and fix edge cases
- Move ToolError and generateRequestID from edit.go to util.go so all
  tools in the package share a single definition rather than relying on
  cross-file visibility
- Add panic guard to Registry.Register for empty tool names (programming
  error caught at startup)
- Fix ImageRefPattern to use (?i) flag so uppercase extensions (.PNG,
  .JPG) are matched correctly
- Add file path context to image load errors for easier debugging
- Mark PermissionAllow/Deny/Ask/PermissionResult aliases as Deprecated
  via godoc comments
- Modernize extractBashCommands to use strings.SplitSeq
