+++ internal/app/input/input.go
@@ -1,6 +1,7 @@
 package input
 
 import (
+	"fmt"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -15,8 +16,8 @@ const (
 	maxTextareaHeight = 6
 )
 
-// ImageRefPattern matches @path/to/image.ext references.
-var ImageRefPattern = regexp.MustCompile(`@([^\s]+\.(png|jpg|jpeg|gif|webp))`)
+// ImageRefPattern matches @path/to/image.ext references (case-insensitive extension).
+var ImageRefPattern = regexp.MustCompile(`(?i)@([^\s]+\.(png|jpg|jpeg|gif|webp))`)
 
 // UpdateHeight adjusts textarea height based on content line count.
 func (m *Model) UpdateHeight() {
@@ -92,7 +93,7 @@ func ProcessImageRefs(cwd, input string) (string, []message.ImageData, error) {
 
 		imgInfo, err := image.Load(absPath)
 		if err != nil {
-			return "", nil, err
+			return "", nil, fmt.Errorf("loading image %s: %w", absPath, err)
 		}
 		images = append(images, imgInfo.ToProviderData())
 		loadedRefs = append(loadedRefs, match[0]) // full match including @
